[BE] Refactor: 보관함 플레이리스트 서비스 분리 및 적용
